+++ src/synthesizer/extensions/grid_props.h
@@ -12,6 +12,7 @@
 #include <Python.h>
 
 /* Local includes */
+#include "data_types.h"
 #include "property_funcs.h"
 
 #pragma omp declare target
@@ -49,15 +50,15 @@ public:
   std::array<int, MAX_GRID_NDIM + 1> unravel_spectra_index(int index) const;
 
   /* Prototypes for getters. */
-  double *get_spectra() const;
-  double get_spectra_at(int grid_ind, int ilam) const;
-  double *get_lam() const;
-  double *get_axis(int idim) const;
-  std::array<double *, MAX_GRID_NDIM> get_all_axes() const;
-  double get_axis_at(int idim, int ind) const;
-  double *get_grid_weights();
+  FLOAT *get_spectra() const;
+  FLOAT get_spectra_at(int grid_ind, int ilam) const;
+  FLOAT *get_lam() const;
+  FLOAT *get_axis(int idim) const;
+  std::array<FLOAT *, MAX_GRID_NDIM> get_all_axes() const;
+  FLOAT get_axis_at(int idim, int ind) const;
+  FLOAT *get_grid_weights();
   PyArrayObject *get_np_grid_weights() const;
-  double get_grid_weight_at(int ind) const;
+  FLOAT get_grid_weight_at(int ind) const;
 
   /* Is wavelength masked? */
   bool lam_is_masked(int ind) const;
@@ -85,7 +86,7 @@ private:
   PyArrayObject *np_grid_weights_;
 
   /* A pointer to the grid weights array data. */
-  double *grid_weights_ = nullptr;
+  FLOAT *grid_weights_ = nullptr;
 
   /* Flag for whether we need to populate the grid weights */
   bool need_grid_weights_ = true;
